Fixed bug with the LED's screwing every 3rd time
The LEDs actually were updated in bunches of three. This had to do with
the fact that our table is 24 cells long instead of 12. Our modulus
however was only for 2, therefore causing lots of mind boggling errors. We
fixed main.c to incorporate for these errors.

Now the i is checked against a modulus of 3. If it is true, the cell needs
to be skipped (because we already wrote to that cell in the previous
iteration). Otherwise, every odd cell needs to be populated with
i-1=0xff and i=0xf0 and every even cell should be i-1=0x0F and i=0xFF. The
last iteration needs to be skipped because of this 3-divisable nature.
Therefore when i becomes 24, it is changed to 1 to start again.

Added some extra comments as well