Changed the coordinate system and get rid of stupid ceiling tricks.
The coordinate system is now the following: X and Y axes represent
coordinates on a single layer where X is the least significant in
gs_buf ordering. Z axis chooses the layer and is the most significant
value in gs_buf.

The nature of TLC5940 makes defining "perfectly non-padded" layers
impossible with layers not dividable by 16. Therefore a new constant
is added: BYTES_PER_LAYER which tells that. There are some
preprocessor macros in cube.c which check if the values in env.h make
any sense.