Generate playlists.c
The idea is that this idea contains a "master" playlist that is then iterated
by the application main. There is also data that may be used to figure out
where the actual playlists reside on that list. This may be used to change
active playlist (handy for serial communication for example).